requirements.txt
flask
pandas
soundfile
librosa
google-generativeai
python-dotenv